+++ lib/cmdlang_stdlib_runner/src/parser_state.ml
@@ -41,11 +41,34 @@ let make_key ~name =
   else "--" ^ name
 ;;
 
+module Arg_presence = struct
+  type 'a t =
+    | Required
+    | Optional
+    | Repeated
+    | With_default of
+        { param : 'a Ast.Param.t
+        ; default : 'a
+        }
+end
+
 let ( let* ) = Result.bind
 
 let make_docv param ~docv =
   let docv = Param_parser.docv param ~docv in
   Printf.sprintf "<%s>" docv
+;;
+
+let make_doc (type a) ~doc ~arg_presence =
+  Printf.sprintf
+    "%s (%s)"
+    doc
+    (match (arg_presence : a Arg_presence.t) with
+     | Required -> "required"
+     | Optional -> "optional"
+     | Repeated -> "repeated"
+     | With_default { param; default } ->
+       Printf.sprintf "default %s" (Param_parser.print param default))
 ;;
 
 let compile
@@ -57,7 +80,8 @@ let compile
   let r = ref [] in
   let pos_state = ref [] in
   let pos_all_state = ref None in
-  let emit s = r := s :: !r in
+  let emit_named s = r := s :: !r in
+  let emit_pos pos = pos_state := Positional_state.One_pos.T pos :: !pos_state in
   let rec aux : type a. a Arg_state.t -> unit =
     fun t ->
     match t with
@@ -70,57 +94,61 @@ let compile
       aux f;
       aux x
     | Flag { names = hd :: tl; doc; var } ->
-      hd :: tl |> List.iter (fun name -> emit (make_key ~name, Arg.Set var, " " ^ doc))
+      let doc = make_doc ~doc ~arg_presence:Optional in
+      hd :: tl
+      |> List.iter (fun name -> emit_named (make_key ~name, Arg.Set var, " " ^ doc))
     | Flag_count { names = hd :: tl; doc; var } ->
+      let doc = make_doc ~doc ~arg_presence:Repeated in
       hd :: tl
       |> List.iter (fun name ->
-        emit (make_key ~name, Arg.Unit (fun () -> incr var), " " ^ doc))
+        emit_named (make_key ~name, Arg.Unit (fun () -> incr var), " " ^ doc))
     | Named { names = hd :: tl; param; docv; doc; var } ->
+      let doc = make_doc ~doc ~arg_presence:Required in
       let docv = make_docv param ~docv in
       hd :: tl
       |> List.iter (fun name ->
-        emit
+        emit_named
           ( make_key ~name
           , make_arg_spec ~name param ~with_var:(fun s -> var := Some s)
           , docv ^ " " ^ doc ))
     | Named_multi { names = hd :: tl; param; docv; doc; rev_var } ->
+      let doc = make_doc ~doc ~arg_presence:Repeated in
       let docv = make_docv param ~docv in
       hd :: tl
       |> List.iter (fun name ->
-        emit
+        emit_named
           ( make_key ~name
           , make_arg_spec ~name param ~with_var:(fun s -> rev_var := s :: !rev_var)
           , docv ^ " " ^ doc ))
     | Named_opt { names = hd :: tl; param; docv; doc; var } ->
+      let doc = make_doc ~doc ~arg_presence:Optional in
       let docv = make_docv param ~docv in
       hd :: tl
       |> List.iter (fun name ->
-        emit
+        emit_named
           ( make_key ~name
           , make_arg_spec ~name param ~with_var:(fun s -> var := Some s)
           , docv ^ " " ^ doc ))
-    | Named_with_default { names = hd :: tl; param; default = _; docv; doc; var } ->
+    | Named_with_default { names = hd :: tl; param; default; docv; doc; var } ->
+      let doc = make_doc ~doc ~arg_presence:(With_default { param; default }) in
       let docv = make_docv param ~docv in
       hd :: tl
       |> List.iter (fun name ->
-        emit
+        emit_named
           ( make_key ~name
           , make_arg_spec ~name param ~with_var:(fun s -> var := Some s)
           , docv ^ " " ^ doc ))
     | Pos { pos; param; docv; doc; var } ->
-      pos_state
-      := Positional_state.One_pos.T { pos; param; docv; doc; presence = Required; var }
-         :: !pos_state
+      let doc = make_doc ~doc ~arg_presence:Required in
+      emit_pos { pos; param; docv; doc; var }
     | Pos_opt { pos; param; docv; doc; var } ->
-      pos_state
-      := Positional_state.One_pos.T { pos; param; docv; doc; presence = Optional; var }
-         :: !pos_state
+      let doc = make_doc ~doc ~arg_presence:Optional in
+      emit_pos { pos; param; docv; doc; var }
     | Pos_with_default { pos; param; default; docv; doc; var } ->
-      pos_state
-      := Positional_state.One_pos.T
-           { pos; param; docv; doc; presence = With_default default; var }
-         :: !pos_state
+      let doc = make_doc ~doc ~arg_presence:(With_default { param; default }) in
+      emit_pos { pos; param; docv; doc; var }
     | Pos_all { param; docv; doc; rev_var } ->
+      let doc = make_doc ~doc ~arg_presence:Repeated in
       pos_all_state := Some (Positional_state.Pos_all.T { param; docv; doc; rev_var })
   in
   aux t;