+++ test/cram/doc.t
@@ -13,20 +13,20 @@
   
 
   $ ./main_climate.exe doc --help
-  Usage: ./main_climate.exe doc [OPTIONS]
-         ./main_climate.exe doc [SUBCOMMAND]
-  
   Testing documentation features
   
   This group is dedicated to testing documentation features.
       
   
-  Options:
-   --help, -h   Print help
+  Usage: ./main_climate.exe doc [COMMAND]
+         ./main_climate.exe doc [OPTIONS]
   
-  Subcommands:
-   args-doc-end-with-dots  Args doc end with dots
-   singleton-with-readme  Singleton command with a readme
+  Options:
+    -h, --help  Print help
+  
+  Commands:
+    args-doc-end-with-dots  Args doc end with dots
+    singleton-with-readme   Singleton command with a readme
   
   This is a readme.
   It can be written on multiple lines.
@@ -107,16 +107,16 @@ A singleton command with a readme:
   
 
   $ ./main_climate.exe doc singleton-with-readme --help
-  Usage: ./main_climate.exe doc singleton-with-readme [OPTIONS]
-  
   Singleton command with a readme
   
   This is a readme.
   It can be written on multiple lines.
   
   
+  Usage: ./main_climate.exe doc singleton-with-readme [OPTIONS]
+  
   Options:
-   --help, -h   Print help
+    -h, --help  Print help
 
   $ ./main_cmdliner.exe doc singleton-with-readme --help=plain
   NAME
@@ -170,10 +170,10 @@ A singleton command with a readme:
     -help   Display this list of options
     --help  Display this list of options
 
-Arguments doc created with or without dots at the end. Positional arguments are
-currently not documented in the help output of the base and climate commands,
-but they are in the cmdliner command. In cmdliner, we currently add the trailing
-dot to the documentation string if it is not present.
+Positional arguments are currently not documented in the help output of the base
+commands. Cmdlang recommmand for arguments doc to be created without dots at the
+end. A dot is systematically added when translating to cmdliner since this
+integrates best with its formatting of help pages.
 
   $ ./main_base.exe doc args-doc-end-with-dots --help
   Args doc end with dots
@@ -186,12 +186,16 @@ dot to the documentation string if it is not present.
   
 
   $ ./main_climate.exe doc args-doc-end-with-dots --help
-  Usage: ./main_climate.exe doc args-doc-end-with-dots [OPTIONS] <STRING> <STRING>
-  
   Args doc end with dots
   
+  Usage: ./main_climate.exe doc args-doc-end-with-dots [OPTIONS] <STRING> <STRING>
+  
+  Arguments:
+    <STRING>  The doc for [a] in the code ends with a dot.
+    <STRING>  The doc for [b] doesn't
+  
   Options:
-   --help, -h   Print help
+    -h, --help  Print help
 
   $ ./main_cmdliner.exe doc args-doc-end-with-dots --help=plain
   NAME
@@ -239,7 +243,7 @@ dot to the documentation string if it is not present.
   Args doc end with dots
   
   Arguments:
-    <STRING>  The doc for [a] in the code ends with a dot (required)
+    <STRING>  The doc for [a] in the code ends with a dot. (required)
     <STRING>  The doc for [b] doesn't (required)
   
   Options: