+++ test/cram/main-help.t
@@ -10,6 +10,7 @@ the executable for each backend.
   
     basic                      . Basic types
     doc                        . Testing documentation features
+    enum                       . Enum types
     named                      . Named arguments
     return                     . An empty command
     version                    . print version information
@@ -17,22 +18,23 @@ the executable for each backend.
   
 
   $ ./main_climate.exe --help
-  Usage: ./main_climate.exe [OPTIONS]
-         ./main_climate.exe [SUBCOMMAND]
-  
   Cram Test Command
   
-  Options:
-   --help, -h   Print help
+  Usage: ./main_climate.exe [COMMAND]
+         ./main_climate.exe [OPTIONS]
   
-  Subcommands:
-   basic  Basic types
-   doc  Testing documentation features
+  Options:
+    -h, --help  Print help
+  
+  Commands:
+    basic   Basic types
+    enum    Enum types
+    doc     Testing documentation features
   
   This group is dedicated to testing documentation features.
       
-   named  Named arguments
-   return  An empty command
+    named   Named arguments
+    return  An empty command
 
   $ ./main_cmdliner.exe --help=plain
   NAME
@@ -47,6 +49,9 @@ the executable for each backend.
   
          doc COMMAND …
              Testing documentation features
+  
+         enum COMMAND …
+             Enum types
   
          named COMMAND …
              Named arguments
@@ -82,6 +87,7 @@ the executable for each backend.
   
   Subcommands:
     basic      Basic types
+    enum       Enum types
     doc        Testing documentation features
     named      Named arguments
     return     An empty command